added protection for which host syncs

+++ sync2drive.sh
@@ -4,7 +4,11 @@ EXTERNAL="/Volumes/SSD"
 
 rsync -va --delete "$HOME/Dropbox/" "$EXTERNAL/share/backups/abiheiri/Dropbox/"
 rsync -va --delete "$HOME/Dropbox/music production/" "$EXTERNAL/share/backups/audio-projects/music production/"
-rsync -va --delete "$HOME/Documents/" "$EXTERNAL/share/backups/abiheiri/twofah docs/"
+
+if [ $(hostname) = "Als-MacBook.local" ];
+then
+	rsync -va --delete "$HOME/Documents/" "$EXTERNAL/share/backups/abiheiri/twofah docs/"
+fi
 
 #Dont delete remove data
 rsync -va "$HOME/Pictures/AutoImport/" "$EXTERNAL/share/backups/backup-pics-videos/AutoImport/"
